requirements.txt
whois
requests
colorama
google-generativeai

